feat(RHINENG-25233): add separate file for topics table assets
